# api-swagger-laravel1/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API Documentation",
        "contact": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "API Documentation"
        }
    ],
    "paths": {
        "/mahasiswa": {
            "get": {
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Mahasiswa - Get All",
                "description": "Mengambil Data Mahasiswa",
                "operationId": "listMahasiswa",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Berhasil Mengambil Data Mahasiswa",
                                    "data": [
                                        {
                                            "mhs_nim": "351610130604",
                                            "mhs_nama": "Helya Salsabilla"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/mahasiswa/{id}": {
            "get": {
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Mahasiswa - Get By ID",
                "description": "Mengambil Data Mahasiswa Berdasarkan ID",
                "operationId": "listMahasiswaById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Mahasiswa",
                        "required": true,
                        "schema": {
                            "description": "Masukan ID",
                            "type": "string",
                            "default": "1",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Berhasil mengambil Data Mahasiswa",
                                    "data": {
                                        "mhs_nim": "0320230010",
                                        "mhs_nama": "Eko Abdul Goffar"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/mahasiswa/create": {
            "post": {
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Mahasiswa Insert",
                "description": "Post data Mahasiswa",
                "operationId": "insertMahasiswa",
                "requestBody": {
                    "description": "Request Body Description",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "mhs_nim": "0320230012",
                                "mhs_nama": "Roni"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Berhasil menambahkan Data Mahasiswa",
                                    "data": {
                                        "mhs_nim": "0320230012",
                                        "mhs_nama": "Roni"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/mahasiswa/update": {
            "put": {
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Mahasiswa - Update",
                "description": "Update data Mahasiswa",
                "operationId": "updateMahasiswa",
                "requestBody": {
                    "description": "Request Body Description",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "mhs_nim": "0320230012",
                                "mhs_nama": "Roni S"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Data berhasil diubah"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/mahasiswa/delete": {
            "delete": {
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Mahasiswa - Delete By ID",
                "description": "Delete data Mahasiswa",
                "operationId": "deleteMahasiswa",
                "requestBody": {
                    "description": "Request Body Description",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "mhs_nim": 320230012
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Berhasil menghapus Data Mahasiswa"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter Token",
                "scheme": "bearer"
            }
        }
    }
}